MapIdentify/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.0
Django==2.1.4
djangorestframework==3.9.0
et-xmlfile==1.0.1
future==0.17.1
geocoder==1.38.1
idna==2.8
include==0.2.2
jdcal==1.4
numpy==1.15.4
openpyxl==2.5.12
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
ratelim==0.1.6
requests==2.21.0
six==1.12.0
urllib3==1.24.1
xlrd==1.2.0
